[refactor]: Extract `style` module
